Created a single pub sub node named thermostat_house
Started with one thermostat which controls the temperature for the whole house.
Create one node named  /thermostat_house . This node is responsible for simulating a
thermostat’s operation in the whole house.

1. Generated a random temperature (between 0°C and 40°C) and store this value in your
class as the current temperature.
For this assignment, temperatures should be represented as integer values.
2. Publish the randomly generated temperature data to t temperature at a frequency
of 3 Hz.
3. In the same node, use a subscriber to listen to t temperature and displays the
received message as Current temperature: X where X is the current temperature
from step 1.