added mode parameter
Added Mode parameter and implemented the logic
away – When this mode is used, the thermostat is turned off. This is the default mode.
eco – When used, it reduces the target temperature by 3 degrees to save energy.
 night – When used, it automatically adjusts the target temperature to a preset comfortable sleeping temperature at night (18°C).

When in away mode, the publisher stops publishing.
In eco mode, the target temperature decreases by 3 degrees. Adjust the current temperature up or down by 1 degree and continue publishing until the target temperature is achieved. Do not go
below 0°C or above 40°C.
 While in night mode, the target temperature is established at 18°C. Alter the current temperature
by 1 degree, either lowering or raising it, and keep publishing until reaching the target
temperature. Do not go below 0°C or above 40°C.